requirements.txt
flask==3.0.0
twilio==8.10.0
python-telegram-bot==21.7
python-dotenv==1.0.0
gunicorn==21.2.0
asgiref==3.8.1
uvicorn==0.32.1
